Revert "Yet to add XMLs"
This reverts commit 5bf670fb9c7be84683b63548824874ae1e462176.
