--- app/src/main/java/org/pragyan/dalal18/fragment/mortgage/MainMortgageFragment.kt
@@ -1,35 +1,0 @@
-package org.pragyan.dalal18.fragment.mortgage
-
-import android.graphics.Color
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.ContextCompat
-import androidx.fragment.app.Fragment
-import androidx.viewpager.widget.ViewPager
-import com.google.android.material.tabs.TabLayout
-import kotlinx.android.synthetic.main.fragment_main_mortgage.*
-import org.pragyan.dalal18.R
-import org.pragyan.dalal18.adapter.MortgagePagerAdapter
-
-class MainMortgageFragment : Fragment() {
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val customView = inflater.inflate(R.layout.fragment_main_mortgage, container, false)
-
-        mortgageViewPager.adapter = MortgagePagerAdapter(childFragmentManager)
-        mainMortgageTabLayout.setupWithViewPager(mortgageViewPager)
-
-        mainMortgageTabLayout.setTabTextColors(ContextCompat.getColor(context!!, R.color.neutral_font_color), ContextCompat.getColor(context!!, R.color.neon_blue))
-        mainMortgageTabLayout.setBackgroundColor(Color.parseColor("#20202C"))
-
-        return customView
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        (activity as AppCompatActivity).supportActionBar?.title = resources.getString(R.string.mortgage_stocks)
-    }
-}
