--- app/src/main/res/layout/fragment_main_mortgage.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <com.google.android.material.tabs.TabLayout
-        android:id="@+id/mainMortgageTabLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="#20202C"
-        android:elevation="8dp"
-        app:layout_constraintBottom_toTopOf="@+id/mortgageViewPager"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:tabGravity="fill"
-        app:tabTextAppearance="@style/TabHeadingStyle" />
-
-    <androidx.viewpager.widget.ViewPager
-        android:id="@+id/mortgageViewPager"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_below="@+id/mainMortgageTabLayout"
-        android:layout_alignParentStart="true"
-        android:background="@color/black_background"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/mainMortgageTabLayout" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
