+++ app/src/main/res/navigation/dalal_navigation.xml
@@ -22,10 +22,10 @@
     </fragment>
 
     <fragment
-        android:id="@+id/main_mortgage_dest"
-        android:name="org.pragyan.dalal18.fragment.mortgage.MainMortgageFragment"
+        android:id="@+id/mortgage_dest"
+        android:name="org.pragyan.dalal18.fragment.MortgageFragment"
         android:label="@string/mortgage"
-        tools:layout="@layout/fragment_main_mortgage" />
+        tools:layout="@layout/fragment_mortgage" />
 
     <fragment
         android:id="@+id/companies_dest"
